+++ devicemanagement/tests/integration/IntegrationTestDevMgmtApiTraceCmds.cpp
@@ -76,6 +76,15 @@ TEST_F(IntegrationTestDevMgmtApiTraceCmds, getSPStatsTraceBuffer) {
   }
 }
 
+TEST_F(IntegrationTestDevMgmtApiTraceCmds, dmStatsRunControl) {
+  if (targetInList({Target::Silicon})) {
+    dmStatsRunControl(false /* Multiple Devices */);
+  } else {
+    DV_LOG(INFO) << "Skipping the test since its not supported on current target";
+    FLAGS_enable_trace_dump = false;
+  }
+}
+
 int main(int argc, char** argv) {
   google::InitGoogleLogging(argv[0]);
   google::SetCommandLineOption("GLOG_minloglevel", "0");